// projects/customer/src/domain/model/customer.model.ts
export interface ICustomerResponse {
  id: number;
  name: string;
  lastName: string;
  email: string;
  phone: string;
  address: string;
  isFrequent: boolean;
}

export interface ICustomer {
  name: string;
  lastName: string;
  email: string;
  phone: string;
  address: string;
  isFrequent: boolean;
}